banco_api/pyproject.toml
[project]
name = "banco-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Vanessa Bustamante",email = "vanobustamante@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "sqlalchemy (<2.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[tool.poetry]
packages = [{include = "banco_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
